+++ kernel/include/exam_segment.h
@@ -4,18 +4,24 @@
 #include <segmem.h>
 #include <string.h>
 
-#define c0_sel  gdt_krn_seg_sel(1)
-#define d0_sel  gdt_krn_seg_sel(2)
-#define c3_sel  gdt_usr_seg_sel(3)
-#define d3_sel  gdt_usr_seg_sel(4)
-#define ts_sel  gdt_krn_seg_sel(5)
+#define c0_idx  1
+#define d0_idx  2
+#define c3_idx  3
+#define d3_idx  4
+#define ts_idx  5
+
+#define c0_sel  gdt_krn_seg_sel(c0_idx)
+#define d0_sel  gdt_krn_seg_sel(d0_idx)
+#define c3_sel  gdt_usr_seg_sel(c3_idx)
+#define d3_sel  gdt_usr_seg_sel(d3_idx)
+#define ts_sel  gdt_krn_seg_sel(ts_idx)
 
 #define address_GDT 0x10000
 #define address_TSS 0x400000
 
-void display_gdt(void);
-seg_desc_t create_dsc_flat(uint64_t type, uint64_t dpl);
-seg_desc_t create_dsc_tss(uint64_t base,uint64_t limit);
+// void display_gdt(void);
+// seg_desc_t create_dsc_flat(uint64_t type, uint64_t dpl);
+// seg_desc_t create_dsc_tss(uint64_t base,uint64_t limit);
 void init_gdt();
 
 #endif
